Tuyet nhat

+++ DoubleClick/database/migrations/2024_12_18_103837_create_loaisach_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::create('loaisach', function (Blueprint $table) {
-            $table->integer('MaLoai')->primary();
+            $table->bigIncrements('MaLoai')->primary();
             $table->string('TenLoai', 30);
             $table->string('SlugLoai', 100);
             $table->string('MoTa', 100)->nullable();
